+++ libs/render_delegate/procedural_custom.cpp
@@ -75,8 +75,8 @@ void HdArnoldProceduralCustom::Sync(HdSceneDelegate *delegate,
                 //ConvertPrimvarToBuiltinParameter(_node, TfToken(paramName.c_str()), p.second.value, &_visibilityFlags, &_sidednessFlags, nullptr);
                 HdArnoldSetParameter(_node, pentry, p.second.value, _renderDelegate);
             } else {
-                // TODO: Should we declare the primvars in that case ?? they are not known by the node but might be used by shaders
-                // if (HdArnoldDeclare(node, TfToken(paramName), const TfToken& scope, const TfToken& type)) {}
+                HdArnoldSetConstantPrimvar(_node, TfToken(paramName), p.second.role, p.second.value, nullptr,
+                    nullptr, nullptr, _renderDelegate);
             }
         }
     }